refactored examples to use render package for object bounding boxes rendering.  added render for object tracking trails.  made all object detection demos use the preprocess.resizer to handle automatic scaling between source image and model input tensor sizes

+++ tracker/convert.go
@@ -5,14 +5,15 @@ import "github.com/swdee/go-rknnlite/postprocess"
 // DetectionsToObjects takes a postprocess object detection results and
 // converts it into a tracker object
 func DetectionsToObjects(dets []postprocess.DetectResult) []Object {
+
 	var objs []Object
 
 	for _, det := range dets {
 
-		x := det.Box.Left
-		y := det.Box.Top
-		width := det.Box.Right - det.Box.Left
-		height := det.Box.Bottom - det.Box.Top
+		x := float32(det.Box.Left)
+		y := float32(det.Box.Top)
+		width := float32(det.Box.Right - det.Box.Left)
+		height := float32(det.Box.Bottom - det.Box.Top)
 
 		objs = append(objs, Object{
 			Rect:  NewRect(float32(x), float32(y), float32(width), float32(height)),
